blackletter/blackletter-upstream/src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.0
python-dotenv==1.0.1

# Database
psycopg==3.1.18
sqlalchemy==2.0.30

# S3 Storage
boto3==1.34.69

# OCR
pypdf==5.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.4.0

# Document Processing
python-docx==1.1.0

# LLM
openai==1.16.0

# Utilities
requests==2.31.0
tqdm==4.66.1
scikit-learn==1.7.1